WIP: Implement source selector filter.
This is more work than I initially anticipated. Seems I never
implemented marshalling of maps and completely skipped over described
types. Filtering based on offset in a quick Event Hubs test works, but
some of the chages are hacky. I want to do a little clean up and think
on some of the approaches before merging.