WIP: Fix race between transfers and detach.
Previously possible for a transfer frame to be in the session tx channel
awaiting send at the same time the link is closed and detach sent.
Remote side could receive the detach before the transfer which is an
error.

Integration tests seem to be passing reliably now. Still some TODOs to
handle before dogfooding and committing to master.