Default batching to false.
Batching has been enabled by default since the early days of this
library. In retrospect I think this was a mistake. I believe most users
would expect messages to be acknowledged immediately when calling
`Message.Accept()` (evidenced in #127). Requiring explicit opt-in
should lessen the risk of bugs due to the batching behavior.